Animations now work
To get animations work I've split the logic of determining where the row 
is clear away from the logic of actually displaying / moving the tiles.

I feel like there is a better way to handle this. My gut tells me that I 
can do this without the multidimensional array, but I need to do more 
research on that.

Known glitch: When items merge left (the only fully working merge) the 
animation does not play.